TODS optimizations: R1 immutable tree, R2 DCLP closed-form, R3 dead-path detection
R=1: nCr difference formula for immutable support update
R=2: Closed-form Case B (d=1) eliminates 93-97% BK, CSR Case A/C, hybrid bucket+overflow PQ
R=3: LeafDeath detection (95-97% paths skip pathSplit), Hitting Set formalization
     with singleton propagation, subsumption, packing LB, component decomposition,
     recursive unit propagation. V17 (lightweight init), V18 (adaptive lazy cache).
All: double type for support/core values (no overflow at any s), hybrid bucket+set PQ.
Bug fixes: prebuiltIndex for V2-V10, (int) cast overflow in all R=2/R=3 variants.
Includes bench_full.sh for comprehensive server experiments.

+++ src/NucleusDecomposition/NucleusCoreDecompositionVertexCentric.h
@@ -50,7 +50,7 @@ class VertexCentricNucleus {
 public:
     VertexCentricNucleus(int numThreads = 16) : numThreads_(numThreads) {}
     
-    std::vector<std::pair<std::vector<daf::Size>, int>> compute(
+    std::vector<std::pair<std::vector<daf::Size>, double>> compute(
         DynamicGraph<TreeGraphNode>& tree,
         const Graph& edgeGraph,
         DynamicGraphSet<TreeGraphNode>& treeGraphV,
@@ -113,7 +113,7 @@ private:
         std::cout << "Vertices: " << n << std::endl;
     }
     
-    std::vector<std::pair<std::vector<daf::Size>, int>> computeCores(
+    std::vector<std::pair<std::vector<daf::Size>, double>> computeCores(
         DynamicGraph<TreeGraphNode>& tree,
         const Graph& edgeGraph,
         int r, int s) {
@@ -210,7 +210,7 @@ private:
         std::cout << "Processed: " << processed << " cliques" << std::endl;
         
         // 转换输出格式
-        std::vector<std::pair<std::vector<daf::Size>, int>> result;
+        std::vector<std::pair<std::vector<daf::Size>, double>> result;
         for (size_t i = 0; i < cliqueIndex.size(); ++i) {
             auto clique = cliqueIndex.byId(i);
             std::vector<daf::Size> vertices;
